Change the interface for skiptoken
